[SandboxIR][SandboxVectorizer][NFC] Move Region to SandboxVectorizer.
When I first added the Region class I was unclear whether it was
SandboxVectorizer-specific or something more general that could go in
SandboxIR (as evidenced by the fact that I got the Region.h include
guard wrong and didn't have to update it for this commit :)

Now I'm convinced that it belongs in SandboxVectorizer, so I'd like to
move it back. This change does just that:

- Region.h, Region.cpp and RegionTest.cpp are moved into the right
directories for SandboxVectorizer.

- The RegionPass and RegionPassManager classes are moved into Region.h.

- All the other touched files are just mechanical changes to make
  everything work. Mainly updating include paths and qualifying some
  names in tests.
