+++ llvm/include/llvm/SandboxIR/PassManager.h
@@ -211,15 +211,6 @@ public:
   bool runOnFunction(Function &F, const Analyses &A) final;
 };
 
-class RegionPassManager final : public PassManager<RegionPass, RegionPass> {
-public:
-  RegionPassManager(StringRef Name) : PassManager(Name) {}
-  RegionPassManager(StringRef Name, StringRef Pipeline,
-                    CreatePassFunc CreatePass)
-      : PassManager(Name, Pipeline, CreatePass) {}
-  bool runOnRegion(Region &R, const Analyses &A) final;
-};
-
 } // namespace llvm::sandboxir
 
 #endif // LLVM_SANDBOXIR_PASSMANAGER_H
