+++ llvm/unittests/SandboxIR/CMakeLists.txt
@@ -8,7 +8,6 @@ set(LLVM_LINK_COMPONENTS
 add_llvm_unittest(SandboxIRTests
   IntrinsicInstTest.cpp
   PassTest.cpp
-  RegionTest.cpp
   OperatorTest.cpp
   SandboxIRTest.cpp
   TrackerTest.cpp
