+++ llvm/unittests/SandboxIR/PassTest.cpp
@@ -13,7 +13,6 @@
 #include "llvm/SandboxIR/Context.h"
 #include "llvm/SandboxIR/Function.h"
 #include "llvm/SandboxIR/PassManager.h"
-#include "llvm/SandboxIR/Region.h"
 #include "llvm/Support/SourceMgr.h"
 #include "gtest/gtest.h"
 
@@ -87,68 +86,6 @@ define void @foo() {
 #endif
 }
 
-TEST_F(PassTest, RegionPass) {
-  auto *F = parseFunction(R"IR(
-define i8 @foo(i8 %v0, i8 %v1) {
-  %t0 = add i8 %v0, 1
-  %t1 = add i8 %t0, %v1, !sandboxvec !0
-  %t2 = add i8 %t1, %v1, !sandboxvec !0
-  ret i8 %t1
-}
-
-!0 = distinct !{!"sandboxregion"}
-)IR",
-                          "foo");
-
-  class TestPass final : public RegionPass {
-    unsigned &InstCount;
-
-  public:
-    TestPass(unsigned &InstCount)
-        : RegionPass("test-pass"), InstCount(InstCount) {}
-    bool runOnRegion(Region &R, const Analyses &A) final {
-      for ([[maybe_unused]] auto &Inst : R) {
-        ++InstCount;
-      }
-      return false;
-    }
-  };
-  unsigned InstCount = 0;
-  TestPass TPass(InstCount);
-  // Check getName(),
-  EXPECT_EQ(TPass.getName(), "test-pass");
-  // Check runOnRegion();
-  llvm::SmallVector<std::unique_ptr<Region>> Regions =
-      Region::createRegionsFromMD(*F);
-  ASSERT_EQ(Regions.size(), 1u);
-  TPass.runOnRegion(*Regions[0], Analyses::emptyForTesting());
-  EXPECT_EQ(InstCount, 2u);
-#ifndef NDEBUG
-  {
-    // Check print().
-    std::string Buff;
-    llvm::raw_string_ostream SS(Buff);
-    TPass.print(SS);
-    EXPECT_EQ(Buff, "test-pass");
-  }
-  {
-    // Check operator<<().
-    std::string Buff;
-    llvm::raw_string_ostream SS(Buff);
-    SS << TPass;
-    EXPECT_EQ(Buff, "test-pass");
-  }
-  // Check pass name assertions.
-  class TestNamePass final : public RegionPass {
-  public:
-    TestNamePass(llvm::StringRef Name) : RegionPass(Name) {}
-    bool runOnRegion(Region &F, const Analyses &A) { return false; }
-  };
-  EXPECT_DEATH(TestNamePass("white space"), ".*whitespace.*");
-  EXPECT_DEATH(TestNamePass("-dash"), ".*start with.*");
-#endif
-}
-
 TEST_F(PassTest, FunctionPassManager) {
   auto *F = parseFunction(R"IR(
 define void @foo() {
@@ -194,65 +131,6 @@ define void @foo() {
   llvm::raw_string_ostream SS(Buff);
   FPM.print(SS);
   EXPECT_EQ(Buff, "test-fpm(test-pass1,test-pass2)");
-#endif // NDEBUG
-}
-
-TEST_F(PassTest, RegionPassManager) {
-  auto *F = parseFunction(R"IR(
-define i8 @foo(i8 %v0, i8 %v1) {
-  %t0 = add i8 %v0, 1
-  %t1 = add i8 %t0, %v1, !sandboxvec !0
-  %t2 = add i8 %t1, %v1, !sandboxvec !0
-  ret i8 %t1
-}
-
-!0 = distinct !{!"sandboxregion"}
-)IR",
-                          "foo");
-
-  class TestPass1 final : public RegionPass {
-    unsigned &InstCount;
-
-  public:
-    TestPass1(unsigned &InstCount)
-        : RegionPass("test-pass1"), InstCount(InstCount) {}
-    bool runOnRegion(Region &R, const Analyses &A) final {
-      for ([[maybe_unused]] auto &Inst : R)
-        ++InstCount;
-      return false;
-    }
-  };
-  class TestPass2 final : public RegionPass {
-    unsigned &InstCount;
-
-  public:
-    TestPass2(unsigned &InstCount)
-        : RegionPass("test-pass2"), InstCount(InstCount) {}
-    bool runOnRegion(Region &R, const Analyses &A) final {
-      for ([[maybe_unused]] auto &Inst : R)
-        ++InstCount;
-      return false;
-    }
-  };
-  unsigned InstCount1 = 0;
-  unsigned InstCount2 = 0;
-
-  RegionPassManager RPM("test-rpm");
-  RPM.addPass(std::make_unique<TestPass1>(InstCount1));
-  RPM.addPass(std::make_unique<TestPass2>(InstCount2));
-  // Check runOnRegion().
-  llvm::SmallVector<std::unique_ptr<Region>> Regions =
-      Region::createRegionsFromMD(*F);
-  ASSERT_EQ(Regions.size(), 1u);
-  RPM.runOnRegion(*Regions[0], Analyses::emptyForTesting());
-  EXPECT_EQ(InstCount1, 2u);
-  EXPECT_EQ(InstCount2, 2u);
-#ifndef NDEBUG
-  // Check dump().
-  std::string Buff;
-  llvm::raw_string_ostream SS(Buff);
-  RPM.print(SS);
-  EXPECT_EQ(Buff, "test-rpm(test-pass1,test-pass2)");
 #endif // NDEBUG
 }
