+++ llvm/unittests/Transforms/Vectorize/SandboxVectorizer/CMakeLists.txt
@@ -11,7 +11,8 @@ add_llvm_unittest(SandboxVectorizerTests
   DependencyGraphTest.cpp
   IntervalTest.cpp
   LegalityTest.cpp
+  RegionTest.cpp
   SchedulerTest.cpp
-  SeedCollectorTest.cpp	
+  SeedCollectorTest.cpp
   VecUtilsTest.cpp
 )
